requirements.txt
pika==1.3.2
python-dotenv
boto3
requests
botocore